+++ src/main.rs
@@ -1,6 +1,6 @@
 use packs::cli;
-
 pub mod packs;
+
 fn main() {
     cli::run()
 }
