+++ src/packs/parser/ruby/mod.rs
@@ -1,0 +1,1 @@
+pub mod packwerk;
